Add FirebasePlacesDatabase & protocol.
The database queries GeoFire to get the place keys around the given point. The
returned keys are then passed to Firebase, which is queried to get the place
details. I do this this way because, afaict, GeoFire can only return the key
for a location and not additional data.

There is much to clean up. See the TODOs in code and the major issues below:
 * We only query Firebase for *one* of the place keys (I couldn't figure out
how to query Firebase for more than one child key)
 * We don't handle errors very well