Issue #49: Add CategoriesUtil.
This class reads from the Yelp file to return a category to parent mapping and
has getRootCategories which uses this mapping to identify the root categories
for a given category. This is useful because UX gave us criteria for filtering
places based on the root categories and the places from Firebase do not have
their corresponding root categories.

Please spell-check my categories from the document, under the "Location criteria" header:
https://docs.google.com/document/d/101jsPjZZ-a_MgZn1UaAyJouufkxzJqG07IrI_0USj1M/edit