Closes #161 - prevent completion handler from being called twice with unique callbacks.
NOTE: I was unable to reproduce the crash and, in any case, don't know how to
test this code so I don't know if this still works as expected - I fixed it
based on continued compilation.

The crash was caused because the delegate contained a mutable completion
handler which, if `fetchEvents` was called more than once, would be replaced,
having the same system completion handler called multiple times.

I removed the delegate in favor of a callback, which seemed simpler. If we
wanted to continue to use the delegate, we could pass a unique identifier into
the fetchEvents call and return that to the delegate, keeping a list of IDs to
completion handlers, but that seemed complicated.