Issue 178: RM unnecessary categories filtering code.
The UX list of places is exhaustive so whitelisting the places UX said
to show is as good as blacklisting the places they said to hide. The
server already uses the whitelist to query yelp (to return fewer results)
so no actions are necessary on the client to handle the top level
categories they specified.

ty to liuche for pointing this out.