Closes #184: Cache results of Place.travelTimes request.
The advantage over the previous implementation is that we return in-flight
requests too.

This will significantly improve #184 - "Travel times don't load when swiping
quickly (rate limited?)" - but it won't make it perfect. One option for
improvement, probably not worth the effort: be smarter about when we
load travel times (e.g. don't load them if we're
in a fast swipe)

Note: I did not test the location update code but it was simple enough that I
thought I could assume it'd work.