Closes #250: Move Remote Config values from FB to client defaults.
This will ensure the client has the most up-to-date configuration on initial
launch, rather than having to wait for the first config to be pulled down (~30
seconds after launch).

`travel_time_expiration_distance` had an empty value on Firebase so I changed
its value to the client default value.

Additionally, the following keys exist on Firebase, but not on the client:
* event_notification_string_1 & 2
* place_details_event_string_1 & 2
* number_of_event_notification_strings
* number_of_place_details_event_strings
* dummy_number

I have taken no action.

I did not check for keys that exist on the client that do not exist in
Firebase.