Closes #358: Separate travel times cache from Place instances.
I tested by:
* Throwing in a conditional debugger statement in Place.travelTimes - we only
set the deferred once for a single place ID
* Tested scrolling across the 20 places at the hotel and I no longer see
missing travel times. Since we get all the travel times to sort, I'm sure this
would be an issue eventually, however.

---

We add a static mutable dictionary from place ID to cached value instead so the
cache can outlive the Place instances.

I roughly addressed a possible existing concurrency issue. If we feel it may be
a problem later, we can file a follow-up.