Closes #392: Display error message if we can't find location.
If the error is showing and we receive a location event, we will dismiss the
dialog and load the places.

Note that we check if a user has location services enabled already so this case
is if the user can't get a location but has services turned on (e.g. in a
basement, something horribly wrong).