No issue: reduce AppConstants reliance on macros.
Macros aren't type safe and miss some advantages of compilation, e.g. enums
are required to be exhaustive. The only advantage I can think of for macros
here is that they write constants into the code so it's marginally faster, but
I think llvm would be smart enough to see an unchanging value and write those
constants into the code anyway so maybe not.