No issue: append to bundle ID for current location; revise mocked name.
This allows us to continue installing release & debug side-by-side. However,
the test target breaks if we change:
- The bundle ID
- The product name (wtf?)

So it's unclear which build is which on device but if devs run from xcode,
it'll be obvious which build they're running. If it becomes a problem, we can
make the identifiers the same.

I filed #503 as a follow-up to clear this up.