No issue: encapsulate loading overlay a little more.
I tried to move the loading overlay to a separate VC but the code remained more
or less the same as this patch:
- init
- fadeIn/OutMessaging
- TapAgainDelegate
- hide

But using it as a VC added complexity (when I call dismiss, which vc am I
dismissing?).

A more encapsulated implementation might have its own PlacesProvider so it can
handle its own fadeIn/OutMessaging, TapAgainDelegate, and hide, but that ties
the VC closely to the PlacesProvider, which is maybe not worth it.